Create a Nostr module and move the base components to that module.  Add module dependency to net.discovery for future Nostr Podcast searcher.
