+++ net/sync/nostr/README.md
@@ -1,0 +1,3 @@
+# :net:sync:nostr
+
+This module contains the sync backend for the Nostr protocol.
