+++ net/sync/nostr/build.gradle
@@ -1,0 +1,24 @@
+plugins {
+    id("com.android.library")
+}
+apply from: "../../../common.gradle"
+
+android {
+    namespace "de.danoeh.antennapod.net.sync.nostr"
+}
+
+dependencies {
+    implementation project(':net:sync:model')
+
+
+    annotationProcessor "androidx.annotation:annotation:$annotationVersion"
+    implementation "androidx.appcompat:appcompat:$appcompatVersion"
+
+    implementation "com.squareup.okhttp3:okhttp:$okhttpVersion"
+    implementation "org.apache.commons:commons-lang3:$commonslangVersion"
+    implementation "commons-io:commons-io:$commonsioVersion"
+    implementation "io.reactivex.rxjava2:rxandroid:$rxAndroidVersion"
+    implementation "io.reactivex.rxjava2:rxjava:$rxJavaVersion"
+
+    testImplementation "junit:junit:$junitVersion"
+}
