+++ net/sync/nostr/src/main/java/de/danoeh/antennapod/net/sync/nostr/model/NostrFilter.java
@@ -1,4 +1,4 @@
-package de.danoeh.antennapod.core.util.nostr;
+package de.danoeh.antennapod.net.sync.nostr.model;
 
 import androidx.annotation.NonNull;
