+++ net/sync/nostr/src/main/java/de/danoeh/antennapod/net/sync/nostr/model/PodcastMetadataEvent.java
@@ -1,4 +1,4 @@
-package de.danoeh.antennapod.core.util.nostr;
+package de.danoeh.antennapod.net.sync.nostr.model;
 
 import org.json.JSONException;
 import org.json.JSONObject;
