+++ net/sync/nostr/src/main/java/de/danoeh/antennapod/net/sync/nostr/util/Bech32Prefix.java
@@ -1,4 +1,4 @@
-package de.danoeh.antennapod.core.util.nostr;
+package de.danoeh.antennapod.net.sync.nostr.util;
 
 /**
  * Author: tcheeric
